Final production code

+++ controllers/task/task.js
@@ -1,3 +1,4 @@
+import ErrorHandler from "../../middlewares/error.js";
 import { Task } from "../../models/task/task.js";
 
 export const newTask = async (req, res, next) => {
@@ -28,26 +29,37 @@ export const getMyTask = async (req, res, next) => {
 };
 
 export const updateTask = async (req, res, next) => {
-  const task = await Task.findById(req.params.id);
+  try {
+    const task = await Task.findById(req.params.id);
 
-  task.isCompleted = !task.isCompleted;
-  await task.save();
+    if (!task) return next(new ErrorHandler("Task not found", 404));
+    task.isCompleted = !task.isCompleted;
 
-  res.status(200).json({
-    success: true,
-    message: "task updated",
-    task,
-  });
+    await task.save();
+
+    res.status(200).json({
+      success: true,
+      message: "task updated",
+      task,
+    });
+  } catch (error) {
+    next(error);
+  }
 };
 
 export const deleteTask = async (req, res, next) => {
+  try {
     const task = await Task.findById(req.params.id);
-  
+    if (!task) return next(new ErrorHandler("Task not found", 404));
+
     await task.deleteOne();
-  
+
     res.status(200).json({
       success: true,
       message: "task deleted",
-      task
+      task,
     });
-  };
+  } catch (error) {
+    next(error);
+  }
+};
